+++ app/connectors_service/connectors/sources/mysql/__init__.py
@@ -1,0 +1,16 @@
+#
+# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+# or more contributor license agreements. Licensed under the Elastic License 2.0;
+# you may not use this file except in compliance with the Elastic License 2.0.
+#
+from .client import MySQLClient, generate_id, row2doc
+from .datasource import MySqlDataSource
+from .validator import MySQLAdvancedRulesValidator
+
+__all__ = [
+    "MySqlDataSource",
+    "MySQLAdvancedRulesValidator",
+    "MySQLClient",
+    "row2doc",
+    "generate_id",
+]
